+++ public/index.css
@@ -76,6 +76,11 @@ body {
     border-radius: 10px;
     box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
     padding: 20px;
+    transition: transform 0.3s ease-in-out;
+}
+
+.feed-element:hover {
+    transform: scale(1.05);
 }
 
 .feed-element img {
